# Improve performance at ntex fortunes