@fafhrd91 Gain ~5% performance in my i7-7700HQ @ 2.8GHz
### Old result:
```json
{
          "latencyAvg": "353.39us", 
          "latencyMax": "9.02ms", 
          "latencyStdev": "214.34us", 
          "totalRequests": 688911, 
          "startTime": 1592709348, 
          "endTime": 1592709363
}, 
{
          "latencyAvg": "627.53us", 
          "latencyMax": "15.28ms", 
          "latencyStdev": "398.60us", 
          "totalRequests": 786430, 
          "startTime": 1592709365, 
          "endTime": 1592709380
}, 
{
          "latencyAvg": "1.12ms", 
          "latencyMax": "26.11ms", 
          "latencyStdev": "815.75us", 
          "totalRequests": 896048, 
          "startTime": 1592709382, 
          "endTime": 1592709397
}, 
{
          "latencyAvg": "2.13ms", 
          "latencyMax": "20.07ms", 
          "latencyStdev": "1.73ms", 
          "totalRequests": 983991, 
          "startTime": 1592709399, 
          "endTime": 1592709415
}, 
{
          "latencyAvg": "3.83ms", 
          "latencyMax": "29.24ms", 
          "latencyStdev": "2.84ms", 
          "totalRequests": 1057442, 
          "startTime": 1592709417, 
          "endTime": 1592709432
},
{
          "latencyAvg": "6.94ms", 
          "latencyMax": "47.57ms", 
          "latencyStdev": "4.11ms", 
          "totalRequests": 1106248, 
          "startTime": 1592709434, 
          "endTime": 1592709449
}
```

### New results
```json
 {
          "latencyAvg": "336.84us", 
          "latencyMax": "8.60ms", 
          "latencyStdev": "198.18us", 
          "totalRequests": 721340, 
          "startTime": 1592718009, 
          "endTime": 1592718024
}, 
{
          "latencyAvg": "592.72us", 
          "latencyMax": "9.30ms", 
          "latencyStdev": "352.86us", 
          "totalRequests": 827659, 
          "startTime": 1592718026, 
          "endTime": 1592718041
}, 
{
          "latencyAvg": "1.06ms", 
          "latencyMax": "12.49ms", 
          "latencyStdev": "708.47us", 
          "totalRequests": 940365, 
          "startTime": 1592718043, 
          "endTime": 1592718058
}, 
{
          "latencyAvg": "2.04ms", 
          "latencyMax": "22.55ms", 
          "latencyStdev": "1.67ms", 
          "totalRequests": 1031379, 
          "startTime": 1592718060, 
          "endTime": 1592718075
}, 
{
          "latencyAvg": "3.70ms", 
          "latencyMax": "29.77ms", 
          "latencyStdev": "2.78ms", 
          "totalRequests": 1097643, 
          "startTime": 1592718077, 
          "endTime": 1592718092
 },
 {
          "latencyAvg": "6.68ms", 
          "latencyMax": "39.30ms", 
          "latencyStdev": "3.92ms", 
          "totalRequests": 1151051, 
          "startTime": 1592718094, 
          "endTime": 1592718109
}
```
I will do another commit where I will add the buffer reserve logic and avoid Option.