@botika 2020-06-22
With `BytesMut` integrate in template engine runs a little faster:
```json
{
          "latencyAvg": "324.46us", 
          "latencyMax": "8.27ms", 
          "latencyStdev": "192.57us", 
          "totalRequests": 748192, 
          "startTime": 1592815867, 
          "endTime": 1592815882
}, 
{
          "latencyAvg": "567.66us", 
          "latencyMax": "8.75ms", 
          "latencyStdev": "338.21us", 
          "totalRequests": 864023, 
          "startTime": 1592815884, 
          "endTime": 1592815899
}, 
{
          "latencyAvg": "1.04ms", 
          "latencyMax": "14.21ms", 
          "latencyStdev": "725.74us", 
          "totalRequests": 963011, 
          "startTime": 1592815901, 
          "endTime": 1592815916
}, 
{
          "latencyAvg": "2.02ms", 
          "latencyMax": "18.36ms", 
          "latencyStdev": "1.67ms", 
          "totalRequests": 1042912, 
          "startTime": 1592815918, 
          "endTime": 1592815933
}, 
{
          "latencyAvg": "3.67ms", 
          "latencyMax": "31.27ms", 
          "latencyStdev": "2.78ms", 
          "totalRequests": 1110324, 
          "startTime": 1592815935, 
          "endTime": 1592815950
}, 
{
          "latencyAvg": "6.65ms", 
          "latencyMax": "48.33ms", 
          "latencyStdev": "3.97ms", 
          "totalRequests": 1156420, 
          "startTime": 1592815952, 
          "endTime": 1592815967
}
```