@botika 2020-06-22
@fafhrd91 I don't know how. You mean to have 2 `SmallVector`, like before the first commit?