@botika 2020-06-22
### Results at https://github.com/TechEmpower/FrameworkBenchmarks/pull/5789/commits/9828d442a15afcb31e1130d74bec78dcc995c684
```json
{
          "latencyAvg": "337.05us", 
          "latencyMax": "9.22ms", 
          "latencyStdev": "204.36us", 
          "totalRequests": 721050, 
          "startTime": 1592825212, 
          "endTime": 1592825227
        }, 
        {
          "latencyAvg": "599.83us", 
          "latencyMax": "9.59ms", 
          "latencyStdev": "360.05us", 
          "totalRequests": 819893, 
          "startTime": 1592825229, 
          "endTime": 1592825244
        }, 
        {
          "latencyAvg": "1.08ms", 
          "latencyMax": "11.02ms", 
          "latencyStdev": "746.64us", 
          "totalRequests": 932176, 
          "startTime": 1592825246, 
          "endTime": 1592825261
        }, 
        {
          "latencyAvg": "2.05ms", 
          "latencyMax": "23.92ms", 
          "latencyStdev": "1.67ms", 
          "totalRequests": 1025227, 
          "startTime": 1592825263, 
          "endTime": 1592825278
        }, 
        {
          "latencyAvg": "3.70ms", 
          "latencyMax": "29.69ms", 
          "latencyStdev": "2.73ms", 
          "totalRequests": 1093736, 
          "startTime": 1592825280, 
          "endTime": 1592825296
        }, 
        {
          "latencyAvg": "6.70ms", 
          "latencyMax": "35.35ms", 
          "latencyStdev": "3.89ms", 
          "totalRequests": 1143871, 
          "startTime": 1592825298, 
          "endTime": 1592825313
        }
```
I don't know what you mean. If you can do a PR to my branch or I reverse https://github.com/TechEmpower/FrameworkBenchmarks/pull/5789/commits/9828d442a15afcb31e1130d74bec78dcc995c684?