@fafhrd91 2020-06-22
hmm, it is slower. lets use your version without https://github.com/TechEmpower/FrameworkBenchmarks/commit/9828d442a15afcb31e1130d74bec78dcc995c684